// frontend/src/views/FlatListView/Offers.d.ts
export type Offers = {
    offerType: string;
    propertyType: string;
    title: string;
    price: number;
    rent: number;
    caution: number;
    area: number;
    pricePerm2: number;
    roomCount: number;
    photos: Array<string>;
    city: string;
    houseNumber: number;
    street: string;
    apartmentNumber: number;
    description: string;
}